… common/api/presentation-backend.api.md
… common/api/presentation-common.api.md
… common/api/summary/presentation-common.exports.csv
… common/changes/@itwin/presentation-backend/presentation-optimizations-for-full-imodel-read_2023-12-22-09-43.json
… common/changes/@itwin/presentation-common/presentation-optimizations-for-full-imodel-read_2023-12-22-09-43.json
… common/config/rush/browser-approved-packages.json
… common/config/rush/pnpm-lock.yaml
… full-stack-tests/presentation/src/performance/GetElementProperties.test.ts
… presentation/backend/package.json
… presentation/backend/src/presentation-backend/PresentationManagerDetail.ts
… presentation/backend/src/test/ElementPropertiesHelper.test.ts
… presentation/backend/src/test/Helpers.ts
… presentation/backend/src/test/PresentationManager.test.snap
… presentation/backend/src/test/PresentationManager.test.ts
… presentation/common/src/presentation-common/LocalizationHelper.ts
… presentation/common/src/presentation-common/PresentationManagerOptions.ts
… presentation/common/src/presentation-common/content/Item.ts
… presentation/common/src/presentation-common/content/PropertyValueFormatter.ts
… presentation/common/src/test/content/Item.test.snap
… presentation/common/src/test/content/Item.test.ts