Closes https://github.com/iTwin/itwinjs-backlog/issues/976.

imodel-native: https://github.com/iTwin/imodel-native/pull/582

Performance of reading all `bis.GeometricElement` elements' properties:

| iModel | Before the change | After the change |
| --- | --- | --- |
| https://construction.bentley.com/22845570-a681-423c-b107-d9d3445a77ec/4d132ce6-df69-441c-ac39-5edef761e42d/imodel-progress | > 4 hours | ~1:30 h |
| https://qa-connect-imodelhubwebsite.bentley.com/Context/892aa2c9-5be8-4865-9f37-7d4c7e75ebbf/iModel/90b6fa70-c130-40c0-8a76-969f57eaaea4 | ~190 s. | ~105 s. |

Also improved the `PresentationManager.getElementProperties` API for the multiple elements case to accept a custom content parser, in case the caller wants to get results in a format other than the simplified `ElementProperties`.